add segregation between legacy code and explicit objects
