+++ RustPlusApi/Examples/GetStorageMonitorChanges/Program.cs
@@ -5,21 +5,18 @@ using RustPlusApi;
 using static __Constants.ExamplesConst;
 
 var rustPlus = new RustPlus(Ip, Port, PlayerId, PlayerToken);
+const uint entityId = 0;
 
 rustPlus.Connected += async (_, _) =>
 {
-    var message = await rustPlus.GetMapAsync();
+    var message = await rustPlus.GetStorageMonitorInfoAsync(entityId);
 
-    if (!message.IsSuccess) return;
-
-    File.WriteAllBytes("map.jpg", message.Data?.JpgImage!);
-
-    message.Data!.JpgImage = null;
     Console.WriteLine($"Infos:\n{JsonConvert.SerializeObject(message, JsonSettings)}");
+};
 
-    Console.WriteLine($"Image saved under: {Directory.GetCurrentDirectory()}");
-
-    rustPlus.Dispose();
+rustPlus.OnStorageMonitorTriggered += (_, message) =>
+{
+    Console.WriteLine($"StorageMonitor:\n{JsonConvert.SerializeObject(message, JsonSettings)}");
 };
 
 await rustPlus.ConnectAsync();
