+++ RustPlusApi/Examples/Legacy/GetEntityInfoLegacy/Program.cs
@@ -4,11 +4,12 @@ using RustPlusApi;
 
 using static __Constants.ExamplesConst;
 
-var rustPlus = new RustPlus(Ip, Port, PlayerId, PlayerToken);
+var rustPlus = new RustPlusLegacy(Ip, Port, PlayerId, PlayerToken);
+const uint entityId = 0;
 
 rustPlus.Connected += async (_, _) =>
 {
-    var message = await rustPlus.GetInfoAsync();
+    var message = await rustPlus.GetEntityInfoLegacyAsync(entityId);
 
     Console.WriteLine($"Infos:\n{JsonConvert.SerializeObject(message, JsonSettings)}");
